+++ views/admin/motion-list/list_all.php
@@ -58,7 +58,8 @@ echo '<h1>' . Yii::t('admin', 'list_head_title') . '</h1>';
 
 echo $this->render('_list_all_export', [
     'hasProposedProcedures' => $hasProposedProcedures,
-    'hasResponsibilities'   => $hasResponsibilities,
+    'hasResponsibilities' => $hasResponsibilities,
+    'search' => $search,
 ]);
 
 echo '<div class="content" data-antragsgruen-widget="backend/MotionList">';